# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Ziziphus IM 服务 REST API 文档",
        "title": "Ziziphus API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/conversations/{conv_id}/webhooks": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Returns all webhooks configured for the specified conversation.",
                "tags": [
                    "webhooks"
                ],
                "summary": "List webhooks",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Conversation ID",
                        "name": "conv_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api.APIResponse"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Creates a new webhook configuration for the specified conversation.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "webhooks"
                ],
                "summary": "Create webhook",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Conversation ID",
                        "name": "conv_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    }
                }
            }
        },
        "/conversations/{conv_id}/webhooks/{webhook_id}": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Updates the configuration of an existing webhook.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "webhooks"
                ],
                "summary": "Update webhook",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Conversation ID",
                        "name": "conv_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Webhook ID",
                        "name": "webhook_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Deletes a webhook configuration from the conversation.",
                "tags": [
                    "webhooks"
                ],
                "summary": "Delete webhook",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Conversation ID",
                        "name": "conv_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Webhook ID",
                        "name": "webhook_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    }
                }
            }
        },
        "/conversations/{conv_id}/webhooks/{webhook_id}/regenerate-key": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Generates a new API key for the specified webhook. The old key becomes invalid.",
                "tags": [
                    "webhooks"
                ],
                "summary": "Regenerate API key",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Conversation ID",
                        "name": "conv_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Webhook ID",
                        "name": "webhook_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    }
                }
            }
        },
        "/conversations/{conv_id}/webhooks/{webhook_id}/test": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Sends a test message via the specified webhook to verify configuration.",
                "tags": [
                    "webhooks"
                ],
                "summary": "Test webhook",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Conversation ID",
                        "name": "conv_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Webhook ID",
                        "name": "webhook_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.APIResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Bearer token from login/register response",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}